% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{get_data}
\alias{get_data}
\title{Import Happiness report 2019 Data}
\usage{
To use this function by itself, store the DF in a variable.
}
\description{
This function reads the data from a csv-file and returns it as a DataFrame-object.
}
\examples{
df = get_data()
}
